patholensProject/requirements.txt
Django == 5.1.3
djangorestframework
drf-spectacular
numpy
django-import-export
django-cors-headers